SQLAssistant/requirements.txt
langchain==0.1.1
langchain_experimental==0.0.49
langchain_google_genai==0.0.6
mysql_connector_repackaged==0.3.1
pandas==2.2.0
python-dotenv==1.0.0
streamlit==1.30.0
